Tambah kamera HP

+++ remove_all_data.py
@@ -16,22 +16,30 @@ conf.read(option.conf)
 
 cf = dict(conf.items('main'))
 
-if not option.csv_only:
-    for url in cf['url'].strip().splitlines():
-        p = urlparse(url)
-        shop_path = p.path.lstrip('/').split('/')[0]
-        web_name = p.netloc.split('.')[-2]
-        download_dir = '-'.join([web_name, shop_path])
-        download_dir = os.path.join(cf['base_download_dir'], download_dir)
-        if not os.path.exists(download_dir):
-            continue
-        for filename in os.listdir(download_dir):
-            filename = os.path.join(download_dir, filename)
-            print('Hapus', filename)
-            os.remove(filename)
-        print('Hapus', download_dir)
-        os.rmdir(download_dir)
+csv_files = []
+for url in cf['url'].strip().splitlines():
+    p = urlparse(url)
+    shop_path = p.path.lstrip('/').split('/')[0]
+    web_name = p.netloc.split('.')[-2]
+    download_dir = '-'.join([web_name, shop_path])
+    csv_file = download_dir + '.csv'
+    csv_files.append(csv_file)
+    if option.csv_only:
+        continue
+    download_dir = os.path.join(cf['base_download_dir'], download_dir)
+    if not os.path.exists(download_dir):
+        continue
+    for filename in os.listdir(download_dir):
+        filename = os.path.join(download_dir, filename)
+        print('Hapus', filename)
+        os.remove(filename)
+    print('Hapus', download_dir)
+    os.rmdir(download_dir)
 
-for filename in glob(f'{cf["category"]}-*.csv'):
+for filename in csv_files:
+    print('Hapus', filename)
+    os.remove(filename)
+filename = cf['category'] + '.csv'
+if os.path.exists(filename):
     print('Hapus', filename)
     os.remove(filename)
